Add CHANGELOG to clarify release notes
Problem:

Users and contributors cannot easily see precisely what has changed
between each version of the project.

This makes it difficult for users to confidently upgrade or manage the
gem as a dependency in their project.

See http://keepachangelog.com/ for more information about the benefits
and best practices around keeping a changelog.

Solution:

Add `CHANGELOG.md`, based on the format used in
https://github.com/thoughtbot/administrate.